requirements.txt
python-dotenv
schedule
requests
litequeue
langchain
langchain-community
langchain-ollama